requirements.txt
streamlit
streamlit-drawable-canvas
reportlab
Pillow
openpyxl
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2